Moved expiry tests to rotate.go
